move all  stories back to the same file
